Remove default and unknown datastore tests.
Depending on the driver and/or the database, it may be possible for
default values for host, port, and unix socket directory to be known,
which means that the expected instance metric would not include
"default" or "unknown".

Since this behavior is specific to the driver/database/language, it is
best to remove them from the generic cross-agent tests.