JBoss Wildfly throws error on startup
By allowing a specific delay to be set in the yml file, using the new parameter jboss7_jmxService_delay an int that is a child of common, the number of seconds to delay monitoring of the jboss jmx service, the admin can ensure that no error is thrown in the JBoss console during startup.  This change has been implemented successfully at a New Relic customer's enterprise JBoss 7 implementation.