Configure Caffeine caches with a same-thread executor
By default Caffeine uses `ForkJoinPool.commonPool()`. Java 17 and 17.0.1
are impacted by a JDK bug [1] which may cause tasks scheduled on the FJP
not to be executed. That in turn may cause out-of-memory errors.

This change matches the behavior of the Guava caches used prior to the
introduction of Caffeine, which also performed cache maintenance in the
context of calling threads.

[1]: https://bugs.openjdk.java.net/browse/JDK-8274349